+++ src/routes/+layout.server.ts
@@ -1,1 +1,10 @@
 export const prerender = true;
+
+if (import.meta.vitest) {
+    const { describe, it, expect } = import.meta.vitest;
+    describe('dummy', () => {
+        it('dummy', () => {
+            expect(1).toBe(1);
+        });
+    });
+};
