+++ tsconfig.json
@@ -8,7 +8,10 @@
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
-		"strict": true
+		"strict": true,
+		"types": [
+			"vitest/importMeta"
+		  ]
 	}
 	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
 	//
