+++ vite.config.ts
@@ -3,7 +3,10 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
 	plugins: [sveltekit()],
+	define: {
+		'import.meta.vitest': false,
+	  },
 	test: {
-		include: ['src/**/*.{test,spec}.{js,ts}']
+		include: ['src/**/*.{test,spec}.{js,ts}', 'src/routes/**/*.{js,ts}']
 	}
 });
